+++ src/ng-vcl/label/label.component.ts
@@ -1,35 +1,15 @@
-import {
-  Component,
-  Input,
-  Output,
-  EventEmitter,
-  ViewChild,
-  HostBinding
-} from '@angular/core';
+import { Directive, Input, HostBinding } from '@angular/core';
 
-@Component({
-  selector: '[vcl-label], vcl-label',
-  templateUrl: 'label.component.html',
+@Directive({
+  selector: '[vcl-label]',
   host: {
     '[class.vclLabel]': 'true'
   }
 })
 export class LabelComponent {
 
-  @Input('label')
-  label: string | undefined;
-
-  @Input('type')
+  @Input('vcl-label')
   type: 'primary' | 'success' | 'info' | 'warning' | 'error' | undefined;
-
-  /**
-   * TODO(issue) this overwrites the users classes
-   * @link https://github.com/angular/angular/issues/7289
-   */
-  // @HostBinding('class')
-  // get labelClass() {
-  //   return this.type ? 'vcl' + this.type.charAt(0).toUpperCase() + this.type.slice(1) : undefined;
-  // }
 
   @HostBinding('class.vclPrimary')
   get vclPrimary() {